+++ app/src/main/java/com/example/lifeline/ui/home/HomeActivity.kt
@@ -1,0 +1,12 @@
+package com.example.lifeline.ui.home
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import com.example.lifeline.R
+
+class HomeActivity: ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_home)
+    }
+}
