+++ app/src/main/java/com/example/lifeline/ui/signup/SignupActivity.kt
@@ -1,9 +1,12 @@
-package com.example.lifeline.presentation.signup
+package com.example.lifeline.ui.signup
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.example.lifeline.R
 import androidx.fragment.app.Fragment
+import com.example.lifeline.presentation.signup.SignupStep1Fragment
+import com.example.lifeline.presentation.signup.SignupStep2Fragment
+import com.example.lifeline.presentation.signup.SignupStep3Fragment
 
 class SignupActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
