+++ app/src/main/java/com/example/lifeline/ui/signup/fragments/SignupStep1Fragment.kt
@@ -1,4 +1,3 @@
-// SignupStep2Fragment.kt
 package com.example.lifeline.presentation.signup
 
 import android.os.Bundle
@@ -7,19 +6,21 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.example.lifeline.R
+import com.example.lifeline.ui.signup.SignupActivity
 
-class SignupStep2Fragment : Fragment() {
+class SignupStep1Fragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        val view = inflater.inflate(R.layout.fragment_signup_step2, container, false)
+        val view = inflater.inflate(R.layout.fragment_signup_step1, container, false)
 
-        val nextButton = view.findViewById<View>(R.id.btn_register_2)
+        val nextButton = view.findViewById<View>(R.id.btn_register_1)
         nextButton.setOnClickListener {
-            (activity as? SignupActivity)?.goToStep(3)
+            (activity as? SignupActivity)?.goToStep(2)
         }
+
         return view
     }
 }
