+++ app/src/main/java/com/example/lifeline/ui/signup/fragments/SignupStep2Fragment.kt
@@ -1,3 +1,4 @@
+// SignupStep2Fragment.kt
 package com.example.lifeline.presentation.signup
 
 import android.os.Bundle
@@ -6,20 +7,20 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.example.lifeline.R
+import com.example.lifeline.ui.signup.SignupActivity
 
-class SignupStep1Fragment : Fragment() {
+class SignupStep2Fragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        val view = inflater.inflate(R.layout.fragment_signup_step1, container, false)
+        val view = inflater.inflate(R.layout.fragment_signup_step2, container, false)
 
-        val nextButton = view.findViewById<View>(R.id.btn_register_1)
+        val nextButton = view.findViewById<View>(R.id.btn_register_2)
         nextButton.setOnClickListener {
-            (activity as? SignupActivity)?.goToStep(2)
+            (activity as? SignupActivity)?.goToStep(3)
         }
-
         return view
     }
 }
